WIP Invoking other lambda
