+++ lambdas/lambda_triggers_etl/lambda_triggers_etl.py
@@ -9,11 +9,11 @@ function lambda_handler.
 import json
 import logging
 import os
-import urllib
 import boto3
 
 from dpypelines.pipeline.shared.shared_lambda.lambda_utils import (
     get_env_variable,
+    get_s3_object_name,
     handle_error,
     trigger_other_lambda
 )
@@ -57,23 +57,3 @@ def process_s3_object(context, s3_object_name:str):
             """
     logger.error(msg)
     handle_error(context, ValueError(msg))
-
-def get_s3_object_name(context, record) -> str:
-    try:
-        object_key = urllib.parse.unquote_plus(
-                record["s3"]["object"]["key"], encoding="utf-8"
-            )
-        bucket_name = record["s3"]["bucket"]["name"]
-
-        # Determine the folder path by taking everything before the last slash.
-        if "/" not in object_key:
-            raise ValueError(f"Object key {object_key} does not appear to be in a folder structure.")
-
-        object_name = f"{bucket_name}/{object_key}"
-        logger.info(f"S3 object name {object_name}")
-        return object_name
-    except KeyError as err:
-        logger.error(
-                f"KeyError when attempting to get bucket name and object key: {str(err)}"
-            )
-        handle_error(context, err)
